Size et all to iOS/iPadOS.

+++ Source/guitarFineTuneFirstClass.cpp
@@ -598,7 +598,7 @@ float guitarFineTuneFirstClass::scaleToGuitarStringSoundsControlWindow()
 		{
 			scaleNow = bndsScaleVerti;
 		}
-		pEksTabbedComponent->setTabBarDepth(tabBarDepthAndroidIosInHorizontal * scaleNow);
+		pEksTabbedComponent->setTabBarDepth((float)tabBarDepthAndroidIosInHorizontal * scaleNow);
 	}
 	else
 	{
@@ -614,7 +614,7 @@ float guitarFineTuneFirstClass::scaleToGuitarStringSoundsControlWindow()
 		{
 			scaleNow = bndsScaleVerti;
 		}
-		pEksTabbedComponent->setTabBarDepth(tabBarDepthAndroidIosInVertical * scaleNow);
+		pEksTabbedComponent->setTabBarDepth((float)tabBarDepthAndroidIosInVertical * scaleNow);
 	}
 
 	return scaleNow;
@@ -712,7 +712,7 @@ void guitarFineTuneFirstClass::resized()
 						{
 							scaleNow = bndsScaleVerti;
 						}
-						pEksTabbedComponent->setTabBarDepth(tabBarDepthAndroidIosInHorizontal * scaleNow);
+						pEksTabbedComponent->setTabBarDepth((float)tabBarDepthAndroidIosInHorizontal * scaleNow);
 					}
 					else
 					{
@@ -728,7 +728,7 @@ void guitarFineTuneFirstClass::resized()
 						{
 							scaleNow = bndsScaleVerti;
 						}
-						pEksTabbedComponent->setTabBarDepth(tabBarDepthAndroidIosInVertical * scaleNow);
+						pEksTabbedComponent->setTabBarDepth((float)tabBarDepthAndroidIosInVertical * scaleNow);
 					}
 					pGuitarFineTuneLookAndFeel->scaleEksLookAndFeelFonts(scaleNow);
 					pGuitarFineTuneLookAndFeel->scaleAllsliderTextBoxes(scaleNow);
@@ -751,12 +751,12 @@ void guitarFineTuneFirstClass::resized()
 					if (curCompntBnds.getWidth() >= curCompntBnds.getHeight())
 					{
 						// Horizontal
-						pEksTabbedComponent->setTabBarDepth(tabBarDepthAndroidIosInHorizontal * scaleNow);
+						pEksTabbedComponent->setTabBarDepth((float)tabBarDepthAndroidIosInHorizontal * scaleNow);
 					}
 					else
 					{
 						// Vertical
-						pEksTabbedComponent->setTabBarDepth(tabBarDepthAndroidIosInVertical * scaleNow);
+						pEksTabbedComponent->setTabBarDepth((float)tabBarDepthAndroidIosInVertical * scaleNow);
 					}
 					pGuitarFineTuneLookAndFeel->scaleEksLookAndFeelFonts(scaleNow);
 					pGuitarFineTuneLookAndFeel->scaleAllsliderTextBoxes(scaleNow);
